moved loginpage to own folder
