+++ lib/ui/screens/timetable_tab.dart
@@ -8,7 +8,7 @@ import 'package:novynaplo/helpers/misc/capitalize.dart';
 import 'package:novynaplo/helpers/misc/delay.dart';
 import 'package:novynaplo/helpers/networkHelper.dart';
 import 'package:novynaplo/helpers/ui/cardColor/timetableCard.dart';
-import 'package:novynaplo/ui/screens/login_page.dart' as login;
+import 'package:novynaplo/ui/screens/login/login_page.dart' as login;
 import 'package:novynaplo/ui/screens/timetable_detail_tab.dart';
 import 'package:novynaplo/i18n/translationProvider.dart';
 import 'package:flutter_calendar_week/flutter_calendar_week.dart';
