+++ lib/ui/widgets/LoadingSpinner.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:novynaplo/i18n/translationProvider.dart';
-import 'package:novynaplo/ui/screens/login_page.dart' as login;
+import 'package:novynaplo/ui/screens/login/login_page.dart' as login;
 
 class LoadingSpinner extends StatefulWidget {
   @override
