+++ lib/ui/widgets/SchoolSearchList.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:novynaplo/API/requestHandler.dart';
 import 'package:novynaplo/i18n/translationProvider.dart';
-import 'package:novynaplo/ui/screens/login_page.dart' as login;
+import 'package:novynaplo/ui/screens/login/login_page.dart' as login;
 
 var searchList = [];
 bool isFetched = false;
